realtime_back.py: fix string to int conversion
We'll often get a wiops value of 0.00000 or so as wiop in
fio now, but since this is a string and we use int("0.00000"), it
will barf:

ValueError: invalid literal for int() with base 10: '0.00000'

Fix this and use float() to cast the value down to '0.0' which,
and so accomplishes the same goal as the cast to int.

+++ realtime_back.py
@@ -27,7 +27,7 @@ def parse_fio_output(riops, rbw, rlat, wiops, wbw, wlat, job_riops, job_rbw, job
                 job_riops[cur_job].append(int(split[7]))
                 job_rbw[cur_job].append(int(split[6]))
                 job_rlat[cur_job].append(float(split[15]))
-                job_wiops[cur_job].append(int(split[48]))
+                job_wiops[cur_job].append(float(split[48]))
                 job_wbw[cur_job].append(int(split[47]))
                 job_wlat[cur_job].append(float(split[56]))
                 cur_job+=1
